+++ recipes/master.rb
@@ -128,6 +128,15 @@ firewall_rule 'kube_apiserver' do
   end
 end
 
+firewall_rule 'kube_controller_manager' do
+  port node['kubernetes']['controller_manager']['secure_port']
+  protocol :tcp
+  interface node['kubernetes']['interface']
+  command :allow
+  only_if do
+    node['kubernetes']['enable_firewall']
+  end
+end
 
 firewall_rule 'kube_scheduler' do
   port node['kubernetes']['scheduler']['secure_port']